# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "V1.0"
  title: "Infusionsoft REST API"
  termsOfService: "https://keys.developer.infusionsoft.com/apps/tos"
  contact:
    name: "Infusionsoft"
    url: "https://developer.infusionsoft.com/get-support/"
host: "api.infusionsoft.com"
basePath: "/crm/rest/v1"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /hooks:
    get:
      tags:
      - "REST Hooks"
      summary: "List Stored Hook Subscriptions"
      description: "Lists your hook subscriptions."
      operationId: "list_stored_hook_subscriptions"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RestHook"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
    post:
      tags:
      - "REST Hooks"
      summary: "Create a Hook Subscription"
      description: "To receive hooks, you'll first need to subscribe to events one\
        \ at a time *and* individually verify each subscription.\n\nThis operation\
        \ is used to create hook subscriptions. During this process, Infusionsoft\
        \ will attempt to verify your subscription. Continue reading to learn how\
        \ that works.\n\nTo verify or reactivate a hook subscription, Infusionsoft\
        \ will make a `POST` request with a temporary secret to the `hookUrl` you\
        \ provided during creation. The secret is passed as the value of a header\
        \ named `X-Hook-Secret`. Your response to that `POST` *must* have a status\
        \ code of `200` and return the same `X-Hook-Secret` header and value pair.\
        \ Once you've done that, you'll begin receiving hooks. Don't worry if verification\
        \ fails; you can always [request another verification attempt](#!/REST_Hooks/verify_a_hook_subscription).\n\
        \nNB: You will not receive events until the subscription is verified.\n\n\
        If the verification process seems confusing, head over to [RESTHooks.org](http://resthooks.org/docs/security/)\
        \ for more on the concept."
      operationId: "create_a_hook_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "restHookRequest"
        description: "restHookRequest"
        required: true
        schema:
          $ref: "#/definitions/RestHookRequest"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/RestHook"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
  /hooks/event_keys:
    get:
      tags:
      - "REST Hooks"
      summary: "List Hook Event Types"
      description: "### _What are REST Hooks?_\n\n_REST Hooks itself is not a specification,\
        \ it is a collection of patterns that treat webhooks like subscriptions. These\
        \ subscriptions are manipulated via a REST API just like any other resource._\n\
        \n_With most modern REST APIs, polling is the only way to detect changes.\
        \ But it doesn't have to be. On average, 98.5% of polls are wasted. There\
        \ is a better way. REST Hooks enable real-time communication and eliminate\
        \ polling._ -- [RESTHooks.org](http://RESTHooks.org)\n\nInfusionsoft offers\
        \ a growing set of REST Hooks to keep you notified of changes in your data.\
        \ To get started, use [List Hook Event Types](#!/REST_Hooks/list_hook_event_types)\
        \ to discover event types. Event types are identified using _noun.verb dot\
        \ syntax_, e.g.: `contact.add` or `invoice.delete`.\n\nOnce you've chosen\
        \ the events you're interested in, subscribe to them via [Create a Hook Subscription](#!/REST_Hooks/create_a_hook_subscription)."
      operationId: "list_hook_event_types"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              type: "string"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
  /hooks/{key}:
    get:
      tags:
      - "REST Hooks"
      summary: "Retrieve a Hook Subscription"
      description: "Retrieves an existing hook subscription and its status.\nIf your\
        \ hook subscription becomes inactive, you may request an activation attempt\
        \ via [Verify a Hook Subscription](#!/REST_Hooks/verify_a_hook_subscription)."
      operationId: "retrieve_a_hook_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "key"
        in: "path"
        description: "key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/RestHook"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
    put:
      tags:
      - "REST Hooks"
      summary: "Update a Hook Subscription"
      description: "Modify an existing hook subscription using the provided values."
      operationId: "update_a_hook_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "key"
        in: "path"
        description: "key"
        required: true
        type: "string"
      - in: "body"
        name: "restHookRequest"
        description: "restHookRequest"
        required: true
        schema:
          $ref: "#/definitions/RestHookRequest"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/RestHook"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
    delete:
      tags:
      - "REST Hooks"
      summary: "Delete a Hook Subscription"
      description: "Stop receiving hooks by deleting an existing hook subscription."
      operationId: "delete_a_hook_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "key"
        in: "path"
        description: "key"
        required: true
        type: "string"
      responses:
        204:
          description: "No Content"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
  /hooks/{key}/verify:
    post:
      tags:
      - "REST Hooks"
      summary: "Verify a Hook Subscription"
      description: "This operation is used to verify or reactivate a hook subscription.\n\
        \nTo verify or reactivate a hook subscription, Infusionsoft will make a `POST`\
        \ request with a temporary secret to the `hookUrl` you provided during creation.\
        \ The secret is passed as the value of a header named `X-Hook-Secret`. Your\
        \ response to that `POST` *must* have a status code of `200` and return the\
        \ same `X-Hook-Secret` header and value pair. Once you've done that, you'll\
        \ begin receiving hooks. Don't worry if verification fails; you can always\
        \ [request another verification attempt](#!/REST_Hooks/verify_a_hook_subscription).\n\
        \nNB: You will not receive events until the subscription is verified.\n\n\
        If the verification process seems confusing, head over to [RESTHooks.org](http://resthooks.org/docs/security/)\
        \ for more on the concept."
      operationId: "verify_a_hook_subscription"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "key"
        in: "path"
        description: "key"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/RestHook"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "RESTHooks"
...
definitions:
...
  RestHook:
    type: "object"
    properties:
      eventKey:
        type: "string"
      hookUrl:
        type: "string"
      key:
        type: "string"
      status:
        type: "string"
        enum:
        - "Unverified"
        - "Verified"
        - "Inactive"
    example:
      eventKey: "eventKey"
      hookUrl: "hookUrl"
      key: "key"
      status: "Unverified"
...
  RestHookRequest:
    type: "object"
    properties:
      eventKey:
        type: "string"
      hookUrl:
        type: "string"
...